Update mainDesktop.js
Hi Yin - this is Brian (you sent me your document for this project earlier this week - it's been very helpful!). You may not have noticed, but there are three errors in your dojo build file. One of them is "error: SyntaxError: missing } after function body". Strangly enough, this is caused by the trailing comment at the end of your file, so I have removed it.

I actually think this is a bug in the dojo builder, but removing the comment is easier than waiting for the next dojo release.

The other 2 errors in your build results file are more complicated. I am working on a solution and will make a pull request if I come up with a fix.